% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsglasso.R
\name{adjacency_from_tsgl}
\alias{adjacency_from_tsgl}
\title{This function gets adjacency matrix for \code{tsglasso}
output}
\usage{
adjacency_from_tsgl(estmat)
}
\arguments{
\item{estmat}{\itemize{
\item three dimensional tensor
}}
}
\value{
*\code{adj} returns adjacency matrix
}
\description{
This function gets adjacency matrix for \code{tsglasso}
output
}
