% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsglasso.R
\name{comparetsg}
\alias{comparetsg}
\title{This function compares estimated inverse spectral density matrix
the true inverse spectral density matrix matrix}
\usage{
comparetsg(esttsg, truetsg)
}
\arguments{
\item{esttsg}{\itemize{
\item Estimated inverse spectral density matrix
}}

\item{truetsg}{\itemize{
\item True inverse spectral density matrix
}}
}
\value{
*\code{tpr} returns true positive rate
*\code{fpr} returns false positive rate
*\code{tdr} returns true discovery rate
}
\description{
This function compares estimated inverse spectral density matrix
the true inverse spectral density matrix matrix
}
