% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvar.R
\name{gendiagAR}
\alias{gendiagAR}
\title{Generates AR coefficient matrix as described in Model 4 inDallakyan et. al 2022 (https://www.sciencedirect.com/science/article/abs/pii/S0167947322001372).}
\usage{
gendiagAR(K, p = 3, block = 5)
}
\arguments{
\item{K}{\itemize{
\item dimension of time series
}}

\item{p}{\itemize{
\item number of lags; default is 3
}}

\item{block}{\itemize{
\item block size; default is 5
}}
}
\value{
*\code{A} returns K x p coefficient matrix
}
\description{
Generates AR coefficient matrix as described in Model 4 inDallakyan et. al 2022 (https://www.sciencedirect.com/science/article/abs/pii/S0167947322001372).
}
