% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvar.R
\name{getcoefadj}
\alias{getcoefadj}
\title{Returns adjacency matrix from the AR matrix}
\usage{
getcoefadj(A)
}
\arguments{
\item{A}{coefficient matrix}
}
\value{
\itemize{
\item returns K x K adjacency matrix
}
}
\description{
Returns adjacency matrix from the AR matrix
}
