% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvar.R
\name{msvar}
\alias{msvar}
\title{Modified sparse VAR algorithm as described in Dallakyan et. al 2022 (https://www.sciencedirect.com/science/article/abs/pii/S0167947322001372).
The \code{msVAR()} function heavily relies on the initial code introduced in Davis et. al (2016)
Most of the functions were taken from the code provided in the supplement of the Davis et. al (2016)}
\usage{
msvar(
  dta,
  p.seq = NULL,
  p.ub = NULL,
  d = NULL,
  allPairs = NULL,
  startValueMethod = c("uniYW", "sparseOLS"),
  standardize = FALSE,
  nonZeroAR = NULL,
  Sigma.given = NULL,
  forceAutoRegression = TRUE,
  stage1.showStatus = FALSE,
  stage2.showStatus = FALSE,
  iteMax = 200,
  reltol = 1e-05,
  stage1.info = c("bic", "aic"),
  halfWindowLength = NULL,
  ADMM_ITER = 30,
  lambda = NULL,
  rho = 100,
  alpha = 1,
  rho.flex = FALSE,
  fdr.q = 0.1,
  pen.method = c("relax", "strict"),
  thresh = 0.005,
  ...
)
}
\arguments{
\item{dta}{Time Series Data}

\item{p.seq}{Sequence of lags}

\item{halfWindowLength}{length for halfwindow}

\item{*}{inputs for \code{tsglasso()}}
}
\description{
Modified sparse VAR algorithm as described in Dallakyan et. al 2022 (https://www.sciencedirect.com/science/article/abs/pii/S0167947322001372).
The \code{msVAR()} function heavily relies on the initial code introduced in Davis et. al (2016)
Most of the functions were taken from the code provided in the supplement of the Davis et. al (2016)
}
