% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvar.R
\name{plot_mat}
\alias{plot_mat}
\title{This function plots the coefficient matrix}
\usage{
plot_mat(Mat, main = NULL)
}
\arguments{
\item{Mat}{\itemize{
\item Coefficient matric
}}

\item{main}{\itemize{
\item title
}}
}
\description{
This function plots the coefficient matrix
}
