% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsglasso.R
\name{postprocess}
\alias{postprocess}
\title{Threshold the output of time series graphical lasso}
\usage{
postprocess(X, threshold = 0.01)
}
\arguments{
\item{X}{\itemize{
\item Estimated inverse spectral density matrix
}}

\item{threshold}{\itemize{
\item Threshold value for the inverse spectral density
}}
}
\value{
*\code{X} returns M x p x p inverse spectral density matrix
*\code{thresh} returns thresholding value
}
\description{
Threshold the output of time series graphical lasso
}
