% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvar.R
\name{simulateVAR}
\alias{simulateVAR}
\title{Generates data from the AR coefficient matrix.}
\usage{
simulateVAR(
  coefMatr = NULL,
  intercept,
  size,
  burn = 100,
  Sigma,
  error.dist = c("normal", "t"),
  df = NULL,
  checkCausal = TRUE
)
}
\arguments{
\item{coefMatr}{K x K Coefficient matrix}

\item{intercept}{K x 1 intercept vector}

\item{size}{Number of observations}

\item{burn}{Number of observations to burn; default is 100}

\item{Sigma}{K x K covariance matrix}

\item{error.dist}{error distribution; default is normal}

\item{df}{degrees of freedom, if \code{error.dist} is t distribution}

\item{checkCausal}{checks whether \code{coefMatr} is causal}
}
\value{
*\code{simData} returns \verb{size x K} data matrix
*\code{coefMatr} coefficient matrix
}
\description{
Generates data from the AR coefficient matrix.
}
