% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsglasso.R
\name{tsglasso}
\alias{tsglasso}
\title{Time series graphical lasso algorithm as described in Dallakyan et. al 2022 (https://www.sciencedirect.com/science/article/abs/pii/S0167947322001372)
and Tugnait 2018 (https://ieeexplore.ieee.org/document/8645324)}
\usage{
tsglasso(
  dta,
  lambda,
  halfWindowLength,
  X_init = NULL,
  rho = 100,
  alpha = 1,
  MAX_ITER = 100,
  method = c("strict", "relax"),
  thresh = 1e-05,
  ABSTOL = 1e-05,
  demean = FALSE,
  RELTOL = 1e-04,
  rho.flex = FALSE,
  diag = FALSE,
  smooth = FALSE
)
}
\arguments{
\item{dta}{\itemize{
\item Time Series Data
}}

\item{lambda}{\itemize{
\item penalty parameter
}}

\item{halfWindowLength}{\itemize{
\item length for halfwindow
}}

\item{X_init}{\itemize{
\item Initial inverse spectral density matrix
}}

\item{rho}{\itemize{
\item ADMM convergence tuning parameter; default is 100
}}

\item{alpha}{\itemize{
\item additional relaxation for ADMM; default is 1
}}

\item{MAX_ITER}{\itemize{
\item Max iteration; default is 100
}}

\item{method}{\itemize{
\item ADMM formulation method; default is strict
}}

\item{thresh}{\itemize{
\item Thresholding the inverse spectral density
}}

\item{ABSTOL}{\itemize{
\item Tolerance for convergence
}}

\item{demean}{\itemize{
\item If TRUE, demeanizes data
}}

\item{RELTOL}{\itemize{
\item Tolerance for convergence
}}

\item{rho.flex}{\itemize{
\item If TRUE, rho is chosen adaptivaly
}}

\item{diag}{\itemize{
\item If TRUE, penalizes diagonal elements
}}

\item{smooth}{\itemize{
\item If TRUE, smoothes the sample estimate of periodagram
}}
}
\value{
*\code{X} returns M x p x p inverse spectral density matrix
*\code{est_S} returns M x p x p spectral density matrix
*\code{hat_S} returns M x p x p sample spectral density matrix
*\code{M}     returns number of frequencies
*\code{K}     returns the window size
*\code{history.objval} returns history of the objective value over iterations
*\verb{history.*} returns history about convergence of the ADMM
}
\description{
Time series graphical lasso algorithm as described in Dallakyan et. al 2022 (https://www.sciencedirect.com/science/article/abs/pii/S0167947322001372)
and Tugnait 2018 (https://ieeexplore.ieee.org/document/8645324)
}
